requirements.txt
langfuse==3.8.0
openai==2.6.0
python-dotenv==1.0.0
requests==2.32.5
langchain==1.0.2
langchain-openai==1.0.1
langchain-community==0.4
